pyproject.toml
[project]
name = "test-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = []

[dependency-groups]
api = [
    "celery>=5.5.2",
    "django-celery-beat>=2.8.1",
    "django-filter>=25.1",
    "django-redis>=5.4.0",
    "djangorestframework>=3.16.0",
    "drf-spectacular[sidecar]>=0.28.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
]
